requirements.txt
streamlit==1.32.0
langchain==0.3.0
langchain-community==0.3.0
langchain-groq
langchain-huggingface
langchain-chroma
sentence-transformers
rank_bm25
chromadb
pypdf
python-docx
pytesseract
Pillow
pydantic
python-dotenv
tiktoken
